+++ lib/startback/services.rb
@@ -1,0 +1,11 @@
+module Startback
+  class Services
+    include Startback::Errors
+
+    def initialize(context)
+      @context = context
+    end
+    attr_reader :context
+
+  end # class Services
+end # module Startback